fix: add force-dynamic to pages to fix prerender build failures
The globalPassThroughEnv: ["*"] in turbo.json (PR #61) causes env vars
to reach Next.js during build, making static prerendering fail with
useContext errors from the root layout's TRPCProvider. Mark all pages as
force-dynamic. Client-only pages (sign-in, pool/create) are extracted
into separate component files with server component wrappers since
'use client' pages ignore the dynamic export.
