+++ apps/web/src/app/not-found.tsx
@@ -1,3 +1,5 @@
+export const dynamic = "force-dynamic";
+
 import Link from "next/link";
 
 export default function NotFound() {
