+++ apps/web/src/app/system-admin/page.tsx
@@ -1,3 +1,5 @@
+export const dynamic = "force-dynamic";
+
 import { redirect } from "next/navigation";
 import { createClient } from "@/lib/supabase/server";
 import { createServerCaller } from "@/lib/trpc/server";
